requirements.txt
aioredis==2.0.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
black==23.7.0
cassandra-driver==3.28.0
click==8.1.5
cqlsh==6.1.2
exceptiongroup==1.1.2
fastapi==0.100.0
geomet==0.2.1.post1
h11==0.14.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
pydantic==2.0.3
pydantic_core==2.3.0
redis==4.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.7.1
uvicorn==0.23.0
